requirements.txt
google-genai
rich
textual
python-dotenv
pyyaml
fastapi
uvicorn[standard]
pydantic[email]
sqlalchemy[asyncio]
aiosqlite
jinja2
aiohttp
beautifulsoup4
markdown
openai
fpdf2
python-multipart
python-jose
websockets
pytest
pytest-asyncio
httpx
